feat(cmd/gpud): move "gpud up" to its own directory
